+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - check for reserved words in widget / param names
 - warning for negative grid spacing
+- keyboard shortcut for export selected widget (cmd/ctrl-opt-shift-f)
+- keyboard shortcut for copy to clipboard (cmd/ctrl-opt-c)
 
 ### Changed
 - extensive refactor/reorganization of code